+++ CHANGELOG.md
@@ -2,13 +2,19 @@
 
 All notable changes to the "styco" extension will be documented in this file.
 
+## [0.0.7]
+
+### Bugfixes
+
+- Fixed a bug, where the corresponding closing tag couldn't be found.
+
 ## [0.0.6]
 
 ### Features
 
-- Create a empty styled component if no style prop is found (#6)
-- Command is now executable not only in the line of the tag-name, but also on the lines below (#3)
+- Create a empty styled component if no style prop is found. (#6)
+- Command is now executable not only in the line of the tag-name, but also on the lines below. (#3)
 
 ### Bugfixes
 
-- If Cursor is placed on style-prop in the same line, the style-prop is now found (#5)
+- If Cursor is placed on style-prop in the same line, the style-prop is now found. (#5)